+++ app/src/main/java/com/practice/hanbitlunch/MainActivity.kt
@@ -15,7 +15,6 @@ import com.google.android.gms.auth.api.signin.GoogleSignInClient
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.practice.designsystem.theme.BlindarTheme
 import com.practice.firebase.R
-import com.practice.hanbitlunch.screen.BlindarNavHost
 import com.practice.preferences.PreferencesRepository
 import com.practice.util.LocalActivity
 import dagger.hilt.android.AndroidEntryPoint
